feat: Init TS
