+++ .eslintrc.json
@@ -6,6 +6,8 @@
     "jest": true,
     "node": true
   },
+  "ignorePatterns": ["test", "docs", "build"],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "sourceType": "module",
     "ecmaVersion": 2021,
@@ -16,10 +18,11 @@
   "settings": {
     "noInlineConfig": true
   },
+  "plugins": ["@typescript-eslint"],
   "extends": [
     "eslint:recommended",
-    "plugin:node/recommended",
-    "airbnb-base",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
     "prettier"
   ],
   "rules": {
