+++ src/app.ts
@@ -1,0 +1,29 @@
+import express from 'express';
+import * as swaggerUI from 'swagger-ui-express';
+import * as path from 'path';
+import * as YAML from 'yamljs';
+import { userRouter } from './resources/users/user.router';
+import { boardRouter } from './resources/boards/board.router';
+import { taskRouter } from './resources/tasks/task.router';
+import { Request, Response, NextFunction } from 'express';
+
+const app = express();
+const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
+
+app.use(express.json());
+
+app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
+
+app.use('/', (req: Request, res: Response, next: NextFunction) => {
+  if (req.originalUrl === '/') {
+    res.send('Service is running!');
+    return;
+  }
+  next();
+});
+
+app.use('/users', userRouter);
+app.use('/boards', boardRouter);
+app.use('/boards', taskRouter);
+
+export default app;
